(edit_drop_menu_cmd): fix -Wsign-conversion warning.

+++ src/editor/editmenu.c
@@ -274,7 +274,7 @@ edit_drop_menu_cmd (WDialog * h, int which)
         menubar->is_active = TRUE;
         menubar->is_dropped = (drop_menus != 0);
         if (which >= 0)
-            menubar->selected = which;
+            menubar->selected = (guint) which;
 
         menubar->previous_widget = dlg_get_current_widget_id (h);
         dlg_select_widget (menubar);
